demo: rename is-even service to random service
