--- demos/is-even/README.md
@@ -1,7 +1,0 @@
-# Is Even Service
-
----
-
-> :warning: **Is even service is still in development**
-
----
