+++ demos/random/pkg/frontend/config.go
