+++ demos/random/pkg/random/config.go
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-package iseven
+package random
 
 // Config contains the node configuration.
 type Config struct {
